:contruction: Cleaner routing pattern with "Secure by Default" routes.
