+++ web/src/routes.ts
@@ -1,7 +1,7 @@
 import { authGuard } from "@auth0/auth0-vue"
 import { createRouter, createWebHistory, type RouteRecordRaw } from "vue-router"
-import adminstrationRoutes from "@/modules/administration/router"
-import centreRoutes from "@/modules/centre/routes"
+
+import adminstrationRoutes from "@/administration/administration-routes"
 
 const routes: RouteRecordRaw[] = [
   {
@@ -9,18 +9,126 @@ const routes: RouteRecordRaw[] = [
     component: async () => await import("@/layouts/Default.vue"),
     children: [
       {
-        name: "Dashboard",
+        path: "",
+        redirect: {
+          name: "Dashboard",
+        },
+      },
+      {
         path: "dashboard",
+        name: "Dashboard",
         component: async () => await import("@/modules/home/views/Dashboard.vue"),
       },
       {
         path: "profile",
         component: async () => await import("@/modules/home/views/Profile.vue"),
       },
-      ...adminstrationRoutes,
-      ...centreRoutes,
+
+      {
+        path: "child-care-centres",
+        component: () => import("@/modules/centre/views/CentreList.vue"),
+      },
+      {
+        // TODO: replace centerId with a slug using https://github.com/simov/slugify on center name
+        // also add the appropriate slug column to the centre table and model
+        path: "child-care-centres/:centreId/:fiscalYearSlug?",
+        component: () => import("@/modules/centre/pages/CentreDashboardPage.vue"),
+        props: true,
+        children: [
+          {
+            path: "",
+            name: "CentreDashboardPage",
+            redirect: {
+              name: "CentreDashboardSummaryPage",
+            },
+          },
+          {
+            path: "summary",
+            component: () => import("@/modules/centre/pages/CentreDashboardSummaryPage.vue"),
+            props: (route) => ({
+              centreId: parseInt(route.params.centreId as string),
+              fiscalYearSlug: route.params.fiscalYearSlug,
+            }),
+            children: [
+              {
+                path: "",
+                name: "CentreDashboardSummaryPage",
+                redirect: {
+                  name: "CentreDashboardSummaryReconciliationPage",
+                },
+              },
+              {
+                path: "reconciliation",
+                name: "CentreDashboardSummaryReconciliationPage",
+                component: () =>
+                  import("@/modules/centre/pages/CentreDashboardSummaryReconciliationPage.vue"),
+                props: (route) => ({
+                  centreId: parseInt(route.params.centreId as string),
+                  fiscalYearSlug: route.params.fiscalYearSlug,
+                }),
+              },
+              {
+                path: "payments",
+                name: "CentreDashboardSummaryPaymentsPage",
+                component: () =>
+                  import("@/modules/centre/pages/CentreDashboardSummaryPaymentsPage.vue"),
+                props: (route) => ({
+                  centreId: parseInt(route.params.centreId as string),
+                  fiscalYearSlug: route.params.fiscalYearSlug,
+                }),
+              },
+            ],
+          },
+          {
+            path: "worksheets",
+            name: "CentreDashboardWorksheetsPage",
+            component: () => import("@/modules/centre/pages/CentreDashboardWorksheetsPage.vue"),
+            props: (route) => ({
+              centreId: parseInt(route.params.centreId as string),
+              fiscalYearSlug: route.params.fiscalYearSlug,
+              month: route.params.month,
+            }),
+            children: [
+              {
+                path: ":month",
+                name: "child-care-centres/worksheets/MonthlyWorksheetPage",
+                component: () =>
+                  import("@/pages/child-care-centres/worksheets/MonthlyWorksheetPage.vue"),
+                props: (route) => ({
+                  centreId: parseInt(route.params.centreId as string),
+                  fiscalYearSlug: route.params.fiscalYearSlug,
+                  month: route.params.month,
+                }),
+              },
+            ],
+          },
+          {
+            path: "employees",
+            name: "CentreDashboardEmployeesPage",
+            component: () => import("@/modules/centre/pages/CentreDashboardEmployeesPage.vue"),
+            props: (route) => ({
+              centreId: parseInt(route.params.centreId as string),
+              fiscalYearSlug: route.params.fiscalYearSlug,
+            }),
+            children: [
+              {
+                path: ":month",
+                name: "CentreDashboardEmployeesMonthlyWorksheetPage",
+                component: () =>
+                  import("@/modules/centre/pages/CentreDashboardEmployeesMonthlyWorksheetPage.vue"),
+                props: (route) => ({
+                  centreId: parseInt(route.params.centreId as string),
+                  fiscalYearSlug: route.params.fiscalYearSlug,
+                  month: route.params.month,
+                }),
+              },
+            ],
+          },
+        ],
+      },
     ],
   },
+  ...adminstrationRoutes,
   {
     path: "/sign-in",
     name: "SignInPage",
